uia: rename uia_definitions to uia_defines (similar to win32defines)
